% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{bubble_sort}
\alias{bubble_sort}
\title{Bubble Sort a vector of doubles}
\usage{
bubble_sort(input)
}
\arguments{
\item{input}{A double vector to sort}
}
\value{
a sorted vector of doubles
}
\description{
Demonstrates using Rust to perform a Bubble Sort on a vector of doubles
}
\examples{
bubble_sort(runif(1000))

}
