% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble-sort.R
\name{bubble_sort_r}
\alias{bubble_sort_r}
\title{Bubble Sorting - R Implementation}
\usage{
bubble_sort_r(nums)
}
\arguments{
\item{nums}{numeric vector to sort}
}
\value{
a sorted integer vector
}
\description{
Bubble Sorting - R Implementation
}
