% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{multithreaded_naive_sqrt}
\alias{multithreaded_naive_sqrt}
\title{Perform a sample multithreaded operation}
\usage{
multithreaded_naive_sqrt(f)
}
\arguments{
\item{f}{a double vector to calculate the square root of}
}
\value{
a double vector of square roots
}
\description{
Demonstrates the performance of hand-coded multithreaded tasks in Rust
}
\details{
This function uses native multithreading fron the Rust standard library to
calculate the square root of each number in a slice of floats. The algorithm
is the "Babylonian Method" of calculating square roots.
}
