% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive-sqrt.R
\name{naive_sqrt}
\alias{naive_sqrt}
\title{R version of the Bablyonian Method for calculating square roots}
\usage{
naive_sqrt(n)
}
\arguments{
\item{n}{a double value (vector length 1)}
}
\value{
the square root of n
}
\description{
R version of the Bablyonian Method for calculating square roots
}
