% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{pass_arrow_record_batch}
\alias{pass_arrow_record_batch}
\title{Pass an Arrow RecordBatch back and forth}
\usage{
pass_arrow_record_batch(srb)
}
\arguments{
\item{srb}{a raw vector representing the contents of an IPC buffer}
}
\value{
RecordBatches to be serialized back to an IPC buffer
}
\description{
Demonstrates a strategy for passing an Arrow RecordBatch from/to R
}
\details{
The majority of the work here is being done by the
\href{crate::structs::RecordBatches}{RecordBatches} struct and associated
functions that transform a serialized RecordBatch (raw vector) into
RecordBatches and back to a serialized RecordBatch using the IPC
specification.
}
\examples{
mtcars_rb <- arrow::record_batch(mtcars)
mtcars_out <- arrow::write_to_raw(mtcars_rb, "stream")
mtcars_in_buffer <- as.raw(pass_arrow_record_batch(mtcars_out))
mtcars_in <- arrow::read_ipc_stream(mtcars_in_buffer)

}
