% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{pass_multiple_characters}
\alias{pass_multiple_characters}
\title{Uppercase a character vector in Rust}
\usage{
pass_multiple_characters(s)
}
\arguments{
\item{s}{A character vector to treat as strings in Rust}
}
\value{
toupper(s)
}
\description{
Demonstrates passing a character (string) vector back and forth between R
and Rust, uppercasing the string in Rust to prove it works.
}
\details{
The vector passed to this function must be contain character values. The Rust
function signature is \verb{pass_multiple_characters_fn(s: Vec<String>) -> Vec<String>}.
extendr_api(v0.2.0) current does not support a method (I could find) for
passing character vectors with NA's, so NA handling will need te be managed
on the R side for now.
}
\examples{
pass_multiple_characters(c("hello", "world"))  # c("HELLO", "WORLD")

}
