% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-wrappers.R
\name{pass_multiple_characters_wrapped}
\alias{pass_multiple_characters_wrapped}
\title{Provides \code{NA} handling for character vectors}
\usage{
pass_multiple_characters_wrapped(s = c("a", "b", NA))
}
\arguments{
\item{s}{A character vector to pass to Rust}
}
\value{
toupper(s)
}
\description{
The extendr API doesn't (as far as I can tell) support NA-handling in
character vectors natively. This function emulates that support by replacing
elements in the output with \code{NA} where they are \code{NA} in the input.
}
\examples{
pass_multiple_characters_wrapped(c("a", 7, NA))  # c("A", "7", NA)
}
