% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{pass_multiple_doubles}
\alias{pass_multiple_doubles}
\title{Multiplies a double vector by 10 in Rust}
\usage{
pass_multiple_doubles(f)
}
\arguments{
\item{f}{A double (float) vector to treat as an extendr_api::Real in Rust}
}
\value{
f * 10
}
\description{
Demonstrates passing a double vector back and forth between R and Rust,
multiplies each element by 10 in Rust to prove it works.
}
\details{
The vector passed to this function must be a double vector (default numeric
vector type in R) or an NA_real_. The Rust function signature is
\verb{fn pass_multiple_doubles_fn(f: Real) -> Real}. NA's, NaN's, Inf's, and
-Inf's are handled implicitly by the f64 specification.
}
\examples{
pass_multiple_doubles(c(5.55, NA, NaN, Inf, -Inf))  # c(55.5, NA, NaN, Inf, -Inf)

}
