% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-wrappers.R
\name{pass_multiple_integers_wrapped}
\alias{pass_multiple_integers_wrapped}
\title{Provides better type handling for \code{pass_multiple_integers()}}
\usage{
pass_multiple_integers_wrapped(n = c(5, 7, 9))
}
\arguments{
\item{n}{A numeric vector to pass to Rust}
}
\value{
n * 10
}
\description{
Coerces input values to integers and replaces output vector elements with NA
where appropriate.
}
\examples{
pass_multiple_integers_wrapped(c(3, 5))   # c(30, 50)
pass_multiple_integers_wrapped(c(1, NA))  # c(10, NA)
}
