% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{pass_multiple_logicals}
\alias{pass_multiple_logicals}
\title{Flip a logical vector in Rust}
\usage{
pass_multiple_logicals(b)
}
\arguments{
\item{b}{A logical (boolean) vector to treat as an extender_api::Logical in Rust}
}
\value{
!b
}
\description{
Demonstrates passing a logical (boolean) vector back and forth between R and
Rust, flipping the values in Rust to prove it works.
}
\details{
The vector passed to this function must be contain logical values, the
boolean type in R, or NAs (plain NA is logical). The Rust function
signature is \verb{fn pass_multiple_logicals_fn(b: Logical) -> Logical}, with
NA's being handled as \code{None} in Rust.
}
\examples{
pass_multiple_logicals(c(TRUE, NA, FALSE))  # c(FALSE, NA, TRUE)

}
