% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{pass_multiple_raws}
\alias{pass_multiple_raws}
\title{Left shift a raw vector in Rust}
\usage{
pass_multiple_raws(r)
}
\arguments{
\item{r}{A raw vector to treat as bytes in Rust}
}
\value{
bitwShiftL(r, 1)
}
\description{
Demonstrates passing a raw (byte) vector back and forth between R and
Rust, left shifting the values in Rust to prove it works.
}
\details{
The vector passed to this function must be contain raw values, the
byte type in R. The Rust function signature is
\verb{pass_multiple_raws_fn(r: &[u8]) -> Vec<u8>}. R does not support NAs for
raw types. Values should be passed as a raw vector, from \code{as.raw()}.
}
\examples{
pass_multiple_raws(as.raw(c(2, 4, 8)))  # c(4, 8, 16)

}
