% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{pass_single_character}
\alias{pass_single_character}
\title{Uppercase a character value in Rust}
\usage{
pass_single_character(s)
}
\arguments{
\item{s}{A character value to treat as a string in Rust}
}
\value{
toupper(s)
}
\description{
Demonstrates passing a character (string) value back and forth between R and
Rust, uppercasing the string in Rust to prove it works.
}
\details{
The value passed to this function must be a character (string) or an
NA_character_. The Rust function signature is
\verb{pass_single_character_fn(s: Option<String>) -> Option<String>}, with NA's
being handled as \code{None} in Rust.
}
\examples{
pass_single_character("hello")        # "HELLO"
pass_single_character(NA_character_)  # NA_character_

}
