% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{pass_single_double}
\alias{pass_single_double}
\title{Multiplies a double by 10 in Rust}
\usage{
pass_single_double(f)
}
\arguments{
\item{f}{A double (float) value to treat as an f64 in Rust}
}
\value{
f * 10
}
\description{
Demonstrates passing a double value back and forth between R and Rust,
multiplies it by 10 in Rust to prove it works.
}
\details{
The value passed to this function must be a double, which is the default
numeric type in R, or an NA_real_. The Rust function signature is
\verb{fn pass_single_float_fn(f: Option<f64>) -> Option<f64>}, with NA's being
handled as \code{None} in Rust. The native R double is double-precision floating
point number, equivalent to Rust's \code{f64}.
}
\examples{
pass_single_double(5)         # 50
pass_single_double(5.5)       # 55
pass_single_double(NA_real_)  # NA_real_

}
