% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-wrappers.R
\name{pass_single_integer_wrapped}
\alias{pass_single_integer_wrapped}
\title{Provides better type handling for \code{pass_single_integer()}}
\usage{
pass_single_integer_wrapped(n = 5)
}
\arguments{
\item{n}{A numeric value to pass to Rust}
}
\value{
n * 10
}
\description{
Coerces input to integers and enforces single values before calling the
Rust implementation.
}
\examples{
pass_single_integer_wrapped(NA)  # NA_integer_
pass_single_integer_wrapped(5)   # 50
}
