% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{pass_single_raw}
\alias{pass_single_raw}
\title{Left shift a raw value in Rust}
\usage{
pass_single_raw(r)
}
\arguments{
\item{r}{A raw value to treat as a byte in Rust}
}
\value{
bitwShiftL(r, 1)
}
\description{
Demonstrates passing a raw (byte) value back and forth between R and
Rust, left shifting the value in Rust to prove it works.
}
\details{
The value passed to this function must be a raw, the byte type in R. The
Rust function signature is \verb{fn pass_single_raw_fn(r: u8) -> u8}. R does
not support NAs for raw types. Currently, extendr (v0.2.0) supports passing
in numbers from 0-255 instead of values from \code{as.raw()} as single values.
Values below 0 are coerced to 0, values greater than 255 are coerced to
255.
}
\examples{
pass_single_raw(4)    # 8
pass_single_raw(-10)  # 0
pass_single_raw(300)  # 254

}
