% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rayon_naive_sqrt}
\alias{rayon_naive_sqrt}
\title{Perform a sample multithreaded operation (rayon)}
\usage{
rayon_naive_sqrt(f)
}
\arguments{
\item{f}{a double vector to calculate the square root of}
}
\value{
a double vector of square roots
}
\description{
Demonstrates the performance of multithreaded tasks using rayon in Rust
}
\details{
This function uses the rayon crate to parallelize calculating the square
root of each number in a slice of floats. The algorithm
is the "Babylonian Method" of calculating square roots.
}
