% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rustbind.R
\docType{package}
\name{rustbind}
\alias{rustbind}
\title{An example package demonstrating the use of the extendr crate from R.}
\description{
An example package demonstrating the use of the extendr crate from R.
}
