% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive-sqrt.R
\name{sapply_naive_sqrt}
\alias{sapply_naive_sqrt}
\title{Apply naive_sqrt() to each item in a vector}
\usage{
sapply_naive_sqrt(n)
}
\arguments{
\item{n}{a double value (vector length 1)}
}
\value{
the square roots of elements in n
}
\description{
Uses a base-R apply function to apply the naive_sqrt() algorithm to each
element in turn. Not parallel.
}
